+++ test/Test/ArraySpec.hs
@@ -10,7 +10,7 @@ import Skeletest
 import System.ByteOrder
 import Telescope.Data.Array
 import Telescope.Data.Axes
-import Telescope.Fits.Encoding.DataArray
+import Telescope.Fits.DataArray
 
 
 spec :: Spec
